refactor last fixes

+++ src/main/java/com/goapi/goapi/domain/dto/appServiceobject/userApi/UserApiRequestArgumentDto.java
@@ -6,6 +6,6 @@ import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import java.io.Serializable;
 
-public record UserApiRequestArgumentDto(Long id, @NotBlank String argName,
+public record UserApiRequestArgumentDto(Integer id, @NotBlank String argName,
                                         @NotNull RequestArgumentType requestArgumentType) implements Serializable {
 }
